% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_check.R
\name{check_overwrite}
\alias{check_overwrite}
\title{Asks for user input if overwrite would invalidate .sge_id}
\usage{
check_overwrite(object, .dir)
}
\description{
Asks for user input if overwrite would invalidate .sge_id
}

