% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_grid_apply.R
\name{clean}
\alias{clean}
\title{Removes all files and folders created by \code{setup} and their contents}
\usage{
clean(dir = getwd())
}
\arguments{
\item{dir}{project directory name}
}
\description{
Removes all files and folders created by \code{setup} and their contents
}
\details{
Removes \code{results/*}, \code{SGE_Output/*}, \code{arg_grid.Rdata},
\code{submit}, \code{seeds.Rdata}, and \code{doone.R}.
}

