% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_dgraph.R
\name{collect.dgraph}
\alias{collect.dgraph}
\title{Collect layers, nodes, or tasks from a completed dgraph}
\usage{
\method{collect}{dgraph}(x, dir = getwd(), layer = NULL, node = NULL,
  task = NULL, ...)
}
\arguments{
\item{x}{\code{dgraph object}}

\item{dir}{director}

\item{layer}{which layer to collect}

\item{node}{which node to collect}

\item{task}{which task id to collect}
}
\description{
Collect layers, nodes, or tasks from a completed dgraph
}

