% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_grid_apply.R
\name{collect.gapply}
\alias{collect.gapply}
\title{Collect completed results files from gresults}
\usage{
\method{collect}{gapply}(x, filter = NULL, regex = NULL, sample = NULL,
  dir = getwd(), ...)
}
\arguments{
\item{x}{object from \code{grid_apply}}

\item{filter}{a quoted string or formula filtering jobs in \code{arg_grid} as in \code{filter}.
Results for jobs matching filter are returned.
If \code{NULL} (default), all results are returned.}

\item{regex}{regular expression  matching files in \code{results/}.
If \code{NULL} (default), all results are returned.}

\item{sample}{the number of files in \code{results/} to sample from.
If \code{NULL} (default), all results are returned.}

\item{dir}{directory}

\item{...}{unused}
}
\description{
Provides arguments to return a sample of the results files of a given size,
or to return results from files matching a regular expression.
}
\details{
\code{filter, regex} and \code{sample} are applied to the available results in order.
For example, results are filtered first, a regex is applied, then a sample is taken.
}

