% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_layer.R
\name{control}
\alias{control}
\title{Control for dgrpah}
\usage{
control(.dgraph, .reps = 1, .mc.cores = 1, .tidy = NULL,
  .backend = "sge", .cache = "all", .verbose = 1)
}
\arguments{
\item{.dgraph}{dgraph}

\item{.reps}{number of replications}

\item{.mc.cores}{number of cores}

\item{.tidy}{whether results are tidied}

\item{.backend}{backend to use. Currently only the default "sge" is defined}

\item{.cache}{which layers/nodes to cache. Currently only the default "all" is defined}

\item{.verbose}{verbosity in printing, can be 1, 2, or 3}
}
\description{
Control for dgrpah
}

