% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapply_summary.R
\name{estimate_time}
\alias{estimate_time}
\title{Estimate time for a given number of reps}
\usage{
estimate_time(object, nreps = NULL)
}
\arguments{
\item{object}{gapply object}

\item{nreps}{number of reps to scale to}
}
\description{
Estimate time for a given number of reps
}
