% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_layer.R
\name{expand_grid_dgraph}
\alias{expand_grid_dgraph}
\title{Return the parameter graph implied by the dgraph grid}
\usage{
expand_grid_dgraph(dgraph, layer.id = NULL)
}
\arguments{
\item{dgraph}{dgraph}

\item{layer.id}{id of the layer to expand. If NULL, expands all layers.}
}
\description{
Return the parameter graph implied by the dgraph grid
}

