% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_grid_apply.R
\name{filter_jobs}
\alias{filter_jobs}
\title{Filter a subset of jobs (rows) from argument grid and modify submission script}
\usage{
filter_jobs(object, ..., .mc.cores = 1, .dir = getwd(), .queue = "long",
  .script.name = "doone.R", .job.name = "distributr",
  .out.dir = "SGE_Output", .email.options = "", .email.addr = "",
  .shell = "csh")
}
\arguments{
\item{object}{gapply object}

\item{...}{arguments to \code{select}}

\item{.mc.cores}{number of cores used to run replications in parallel (can be a range)}

\item{.queue}{name of queue}

\item{.script.name}{name of script}

\item{.job.name}{name of job}

\item{.out.dir}{name of directory in which to put SGE output files.}

\item{.email.options}{one or more characters from "bea" meaning email when "job Begins", "job Ends", and "job Aborts". Default is "a".}

\item{.email.addr}{email address}

\item{.shell}{shell to use. Default is 'bash'}
}
\description{
Filter a subset of jobs (rows) from argument grid and modify submission script
}

