% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_layer.R
\name{is.dgraph}
\alias{is.dgraph}
\alias{is.layer}
\alias{is.node}
\title{Checks if x is a dgraph, node or layer}
\usage{
is.dgraph(x)

is.layer(x)

is.node(x)
}
\arguments{
\item{x}{object}
}
\description{
Checks if x is a dgraph, node or layer
}
\section{Functions}{
\itemize{
\item \code{is.layer}: Checks if x is a layer

\item \code{is.node}: Checks if x is a node
}}

