% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_jobs.R
\name{jobs}
\alias{jobs}
\title{Returns the parameter grid from grid_apply}
\usage{
jobs(object)
}
\arguments{
\item{object}{object from grid_apply}
}
\description{
Returns the parameter grid from grid_apply
}

