% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_layer.R
\name{layer}
\alias{grid_map}
\alias{grid_reduce}
\alias{layer}
\title{Add layer to dgraph}
\usage{
layer(..., .reduce = FALSE)

grid_map(..., .reduce = FALSE)

grid_reduce(...)
}
\arguments{
\item{...}{nodes, separated by commas}

\item{.reduce}{if TRUE, the results of previous layers are expanded by nodes, otherwise reduced}
}
\description{
Add layer to dgraph

Alias for \code{layer}
}
\section{Functions}{
\itemize{
\item \code{grid_map}: alias for \code{layer}

\item \code{grid_reduce}: alias for \code{layer(..., .reduce=TRUE)}
}}

