% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_layer.R
\name{layer_apply}
\alias{layer_apply}
\title{Apply function to all nodes in a layer}
\usage{
layer_apply(l, select = ".id", FUN = I)
}
\arguments{
\item{l}{layer}

\item{select}{name of object in layer to select. Default is a node, given by ".id"}

\item{FUN}{function to apply}
}
\description{
Apply function to all nodes in a layer
}

