% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_dgraph.R
\name{load_results}
\alias{load_results}
\title{Load files from \code{results/} satisfying regular expression}
\usage{
load_results(regex, dir = getwd())
}
\arguments{
\item{regex}{regex}

\item{dir}{directory}
}
\value{
returns list of results
}
\description{
Load files from \code{results/} satisfying regular expression
}

