% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_help.R
\name{mysys}
\alias{mysys}
\title{Run and \code{cat} a call to \code{system}}
\usage{
mysys(cmd)
}
\arguments{
\item{cmd}{string of the command to run}
}
\description{
Run and \code{cat} a call to \code{system}
}
