% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_layer.R
\name{node}
\alias{node}
\title{Apply a function to a grid of its parameters in a dgraph}
\usage{
node(.f, ..., .dep = NULL, .id = NULL)
}
\arguments{
\item{.f}{function}

\item{...}{arguments to \code{.f} given as vectors as in \code{grid_apply}}

\item{.dep}{which nodes does this one depnend on?
If NULL, depends on all nodes of previous layer.}

\item{.id}{custom id. if \code{NULL}, one will be assigned.}
}
\description{
Apply a function to a grid of its parameters in a dgraph
}

