% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_monitor.R
\name{parse_qstat}
\alias{parse_qstat}
\title{Parse qst(xml=T) to data frame}
\usage{
parse_qstat(jstr)
}
\arguments{
\item{jstr}{xml string from \code{qst()}}
}
\description{
Parse qst(xml=T) to data frame
}
