% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_gapply.R
\name{plot.gapply}
\alias{plot.gapply}
\title{Plot simulation object}
\usage{
\method{plot}{gapply}(object, .fun = mean, .key = NULL)
}
\description{
Automatic plotting method for aggregating over replications
}

