% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_monitor.R
\name{qst}
\alias{qst}
\title{Return running jobs as a string}
\usage{
qst(xml = TRUE)
}
\arguments{
\item{xml}{whether qstat returns output as xml (default: TRUE)}
}
\description{
Return running jobs as a string
}

