% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_monitor.R
\name{qst_meta}
\alias{qst_meta}
\title{Retrieve meta data for a job id (in xml)}
\usage{
qst_meta(jid, xml = FALSE)
}
\arguments{
\item{jid}{job id}

\item{xml}{whether xml is returned (default: \code{FALSE})}
}
\description{
Retrieve meta data for a job id (in xml)
}

