% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_qstat.R
\name{qstat}
\alias{qstat}
\title{Meta-data for currently running jobs}
\usage{
qstat(user = TRUE)
}
\arguments{
\item{user}{job meta-data is returned for user jobs only (default: \code{TRUE}),
otherwise meta-data for all jobs in default queue is returned}
}
\value{
A data frame with the following columns:
\item{job-ID}{SGE job id}
\item{prior}{Priority of the job}
\item{name}{job name}
\item{user}{user name}
\item{state}{ state of the job, e.g. \code{r} for running}
\item{start}{ date and time of job start}
\item{queue}{queue and machine job is running on}
\item{jclass}{class of job (usually \code{NA})}
\item{slots}{Number of slots (cores) job is running with}
\item{.sge_id}{the value of \code{SGE_TASK_ID} for the job}
}
\description{
Meta-data for currently running jobs
}

