% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_layer.R
\name{reps}
\alias{reps}
\title{Run all computations in a graph repeatedly}
\usage{
reps(.dgraph, .reps)
}
\arguments{
\item{.dgraph}{dgraph}

\item{.reps}{number of replications to run}
}
\description{
Run all computations in a graph repeatedly
}

