% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_grid_apply.R
\name{setup.gresults}
\alias{setup.gresults}
\title{Setup sge files from gresults}
\usage{
\method{setup}{gresults}(object, .dir = getwd(), .reps = 1, .mc.cores = 1,
  .verbose = 1, .queue = "long", .script.name = "doone.R",
  .job.name = "distributr", .out.dir = "SGE_Output", .email.options = "a",
  .email.addr = "", .shell = "csh")
}
\description{
Setup sge files from gresults
}

