% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_grid_apply.R
\name{setup}
\alias{setup}
\title{setup sge simulation}
\usage{
setup(object, dir = getwd(), .reps = 1, .chunks = 1, .mc.cores = 1,
  .verbose = 1, .queue = "long", .script.name = "doone.R",
  .job.name = "patr1ckm", .out.dir = "SGE_Output", .email.options = "a",
  .email.addr = "patr1ckm.crc.nd.edu", .shell = "bash")
}
\arguments{
\item{object}{gapply object}

\item{dir}{directory name relative to the current working directory, ends in '/'}

\item{.reps}{total number of replications for each condition}

\item{.chunks}{split \code{.reps} across this many nodes (see details)}

\item{.mc.cores}{number of cores used to run replications in parallel}

\item{.verbose}{verbose level}

\item{.queue}{name of queue}

\item{.script.name}{name of script}

\item{.job.name}{name of job}

\item{.out.dir}{name of directory in which to put SGE output files.}

\item{.email.options}{one or more characters from "bea" meaning email when "job Begins", "job Ends", and "job Aborts". Default is "a".}

\item{.email.addr}{email address}

\item{.shell}{shell to use. Default is 'bash'}
}
\description{
setup sge simulation
}
\details{
The replications performed per chunk is computed as \code{ceiling(.reps/.chunks)}, which
will produce more total replications than requested if \code{.reps} is not evenly divisible by \code{.chunks}
}

