% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_env.R
\name{sge_env}
\alias{sge_env}
\alias{sge_task_id}
\alias{sge_slots}
\alias{sge_job_id}
\alias{sge_job_name}
\alias{on_sge}
\title{SGE environmental variables}
\usage{
sge_task_id()

sge_slots()

sge_job_id()

sge_job_name()

on_sge()
}
\description{
SGE environmental variables
}
\details{
These functions return environmental variables set by SGE with each job.
}
\section{Functions}{
\itemize{
\item \code{sge_task_id}: returns current \code{SGE_TASK_ID}

\item \code{sge_slots}: returns current number of slots

\item \code{sge_job_id}: returns the current job id

\item \code{sge_job_name}: returns the job name

\item \code{on_sge}: is TRUE if running on SGE
}}

