% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapply_tidy.R
\name{stack_list}
\alias{stack_list}
\title{Stacks a list of vectors, lists, or data frames}
\usage{
stack_list(xl)
}
\arguments{
\item{xl}{list of vectors, lists, or data frames}
}
\value{
tibble with \code{key}, and \code{value} as columns; \code{key2} if list of data frames. \code{key} and \code{key2} are not factors.
}
\description{
Stacks a list of vectors, lists, or data frames into a tibble with \code{key} and \code{value} columns,
us
}
\details{
Stacks a vector or list into \code{key} and \code{value} columns, where \code{key} takes the names of the elements,
 if the names are null, assigns names.

Stacks a list of vectors into \code{key} and \code{value} columns, where \code{key} takes the names of the elements,
 if the names are null, assigns names.

Stacks a data frame or list of vectors of the same length into \code{key}, \code{key2}, and \code{value} columns,
 where \code{key} and \code{key2} are the column and row names of the first element of \code{xl}. If names are null, assigns names.
}
