% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapply_tidy.R
\name{stack_x}
\alias{stack_x}
\title{Stack a simple vector or data frame into key and value columns}
\usage{
stack_x(x)
}
\arguments{
\item{x}{a vector or data frame}
}
\description{
Stacks a vector or list into \code{key} and \code{value} columns, where key takes the names of the elements, or
assigns names if null.
}
\details{
Stacks a list of vectors into \code{key} and \code{value} columns, where key names its elements automatically

Stacks a data frame into \code{key} (key2) and \code{value} columns, where key and key2 are the column
and row names of x.
}
