% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapply_summary.R
\name{summary.gresults}
\alias{summary.gresults}
\title{Print a summary of \code{gresults}}
\usage{
\method{summary}{gresults}(object, .reps = NULL, ...)
}
\arguments{
\item{object}{gresults object}

\item{.reps}{number of reps to scale to}

\item{...}{unused}
}
\value{
Prints the first few conditions, a summary of conditions, and estimated time to scale up to x reps.
 Returns the estimated time invisibly.
}
\description{
Print a summary of \code{gresults}
}

