% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_jobs.R
\name{test_job}
\alias{test_job}
\title{Runs a job on the head node}
\usage{
test_job(.sge_id = 1, .mc.cores = 1, .script.name = "doone.R")
}
\arguments{
\item{.sge_id}{job id (row of \code{jobs(object)})}

\item{.mc.cores}{number of cores to use}

\item{.script.name}{name of the script given in \code{setup}}
}
\description{
Runs a job on the head node
}
\details{
Runs \code{Rscript doone.R x y} at the command line where \code{x} is \code{.sge_id},
and \code{y} is \code{.mc.cores}.

Note that long running jobs claiming many cores may be killed by
the administrators. This should only be used for short tests, longer tests
should be submitted to the cluster as usual.
}
\seealso{
\code{submit}, \code{filter}
}

