% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_results.R
\name{tidy.dgraph}
\alias{tidy.dgraph}
\title{Tidying dgraph results}
\usage{
\method{tidy}{dgraph}(x, arg_grid = NULL, dir = getwd(), layer.id = NULL,
  ...)
}
\arguments{
\item{x}{list of results}

\item{arg_grid}{argument grid}

\item{dir}{directory}

\item{layer.id}{index of the layer to tidy. If \code{NULL} tidies last layer.}

\item{...}{unused}
}
\description{
Tidying dgraph results
}

