% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_apply.R
\name{wrapWE}
\alias{wrapWE}
\title{Wraps a function so that it catches and returns warnings and errors as attributes}
\usage{
wrapWE(fun)
}
\arguments{
\item{fun}{function}
}
\value{
\code{fun}
}
\description{
Wraps a function so that it catches and returns warnings and errors as attributes
}
