% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_jobs.R
\name{add_jobs}
\alias{add_jobs}
\title{Add jobs (or rows) to argument grid}
\usage{
add_jobs(object, ...)
}
\arguments{
\item{object}{grid_apply object}

\item{...}{key value pairs for \code{.f} in \code{grid_apply}}
}
\value{
\code{grid_apply} object with updated
}
\description{
Add jobs (or rows) to argument grid
}
\details{
If all original keys are not in ..., the values of these arguments are set
to \code{NA}
}

