% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_grid_apply.R
\name{collect.gresults}
\alias{collect.gresults}
\title{Collect completed results files from gresults}
\usage{
\method{collect}{gresults}(object, dir = getwd())
}
\description{
Collect completed results files from gresults
}

