% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_layer.R
\name{layer}
\alias{layer}
\title{Add layer to dgraph}
\usage{
layer(..., .reduce = FALSE)
}
\arguments{
\item{...}{nodes, separated by commas}

\item{.reduce}{if TRUE, the results of previous layers are expanded by nodes, otherwise reduced}
}
\description{
Add layer to dgraph
}

