% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_monitor.R
\name{qstat}
\alias{qstat}
\title{Meta-data for currently running jobs}
\usage{
qstat(user = TRUE)
}
\arguments{
\item{user}{job meta-data is returned for user jobs only (default: \code{TRUE}),
otherwise meta-data for all jobs in default queue is returned}
}
\value{
A data frame with the following columns:
\item{job_id}{SGE job id}
\item{.sge_id}{the value of \code{SGE_TASK_ID} for the job}
\item{prior}{Priority of the job}
\item{name}{job name}
\item{user}{user name}
\item{state}{ state of the job, e.g. \code{r} for running}
\item{start}{ date and time of job start}
\item{queue}{queue and machine job is running on}
\item{jclass}{class of job (usually \code{NA})}
\item{slots}{Number of slots (cores) job is running with}
\item{maxvmem}{Maximum virtual memory used for the job (GB)}
\item{mem}{Current physical memory used for the job (GB)}
\item{vmem}{Current virtual memory used for the job (GB)}
\item{wallclock}{Amount of time the job has been running (sec)}
\item{cpu}{Amount of CPU time the job has used (sec)}
The columns \code{prior}, \code{user}, \code{start}, \code{queue}, and \code{jclass}
are not printed. If no jobs are running or in the queue, returns \code{data.frame()}.
For queued jobs, \code{queue = NA}, and \code{.sge_id} is the lowest task
still in the queue. Does not work for jobs with job classes.
}
\description{
The meta-data of running SGE jobs are returned as a data frame. The
meta includes status, resources requested, memory usage, and wallclock/cpu time.
Works for all grid engine jobs and task arrays, including the ones created by \code{distributr}.
}
\examples{
\dontrun{

sge_test(wait=20)
qstat()
}
}

