% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_grid_apply.R
\name{setup}
\alias{setup}
\title{setup sge simulation}
\usage{
setup(x, ...)
}
\arguments{
\item{object}{gapply object}

\item{dir}{directory name relative to the current working directory, ends in '/'}

\item{.reps}{total number of replications for each condition}

\item{.chunks}{split \code{.reps} across this many nodes (see details)}

\item{.mc.cores}{number of cores used to run replications in parallel (can be a range)}

\item{.verbose}{verbose level}

\item{.script.name}{name of script}

\item{.queue}{name of queue}

\item{.job.name}{name of job}

\item{.out.dir}{name of directory in which to put SGE output files.}

\item{.email.options}{one or more characters from "bea" meaning email when "job Begins", "job Ends", and "job Aborts". Default is "a".}

\item{.email.addr}{email address}

\item{.shell}{shell to use. Default is 'bash'}
}
\description{
setup sge simulation
}
\details{
The replications performed per chunk is computed as \code{ceiling(.reps/.chunks)}, which
will produce more total replications than requested if \code{.reps} is not evenly divisible by \code{.chunks}
}

