% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_grid_apply.R
\name{sge_test}
\alias{sge_test}
\title{sge_test
Simple test SGE configuration}
\usage{
sge_test(dir = getwd())
}
\arguments{
\item{dir}{directory in which to run test}
}
\value{
A \code{gresults} object with the submitted plan
}
\description{
sge_test
Simple test SGE configuration
}

