% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_results.R
\name{tidy.dgraph}
\alias{tidy.dgraph}
\title{Tidying dgraph results}
\usage{
\method{tidy}{dgraph}(x, arg_grid = NULL, dir = getwd(), layer.id = NULL)
}
\arguments{
\item{x}{list of results}

\item{dir}{directory}

\item{layer}{layer if not last}
}
\description{
Tidying dgraph results
}

